Update to use AbstractApi superclass
